requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alpaca-py==0.33.0
alpaca-trade-api==3.2.0
alpha_vantage==3.0.0
annotated-types==0.7.0
appdirs==1.4.4
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleak==0.22.2
blinker==1.8.1
ccxt==4.4.32
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
datasets==3.0.1
decorator==5.1.1
deprecation==2.1.0
dill==0.3.8
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.7.0
docker-pycreds==0.4.0
duckdb==1.1.3
email_validator==2.2.0
exchange_calendars==4.5.8
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Principal==0.4.0
Flask-Security==5.5.2
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.0
gTTS==2.5.2
h11==0.14.0
h5py==3.12.1
holidays==0.61
html5lib==1.1
huggingface-hub==0.25.2
ibapi==9.81.1.post1
idna==3.7
ijson==3.3.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jsonpickle==4.0.0
kagglehub==0.3.3
keras==3.7.0
kiwisolver==1.4.7
korean-lunar-calendar==0.3.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lumibot==3.8.9
lumiwealth-tradier==0.1.12
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
MetaTrader5==5.0.4687
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mysqlclient==2.2.4
namex==0.0.8
networkx==3.4
newsapi-python==0.2.7
nltk==3.9.1
nnfs==0.5.1
numba==0.60.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.3
pandas-datareader==0.10.0
pandas_market_calendars==4.4.2
parso==0.8.4
passlib==1.7.4
peewee==3.17.8
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
polygon-api-client==1.14.2
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==17.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.0
Pygments==2.18.0
pyluach==2.2.0
pyparsing==3.2.0
pyspellchecker==0.8.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.1
Quandl==3.7.0
quantstats-lumi==0.3.3
regex==2024.9.11
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.17.0
setproctitle==1.3.3
setuptools==75.1.0
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.36
sqlparse==0.5.1
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow_intel==2.18.0
termcolor==2.5.0
thetadata==0.9.11
threadpoolctl==3.5.0
tokenizers==0.20.1
toolz==1.0.0
torch==2.4.1
torchaudio==2.4.1
torchtext==0.18.0
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
uuid==1.30
wandb==0.18.5
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.2
wget==3.2
wheel==0.45.1
winrt-runtime==2.2.0
winrt-Windows.Devices.Bluetooth==2.2.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.2.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.2.0
winrt-Windows.Devices.Enumeration==2.2.0
winrt-Windows.Foundation==2.2.0
winrt-Windows.Foundation.Collections==2.2.0
winrt-Windows.Storage.Streams==2.2.0
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.2.1
xxhash==3.5.0
yarl==1.15.4
yfinance==0.2.49
youtube-dl==2021.12.17
